requirements.txt
google-generativeai>=0.3.2
pymupdf>=1.23.8
pdfplumber>=0.10.3
streamlit>=1.31.0
chromadb>=0.4.22
python-dotenv>=1.0.0
numpy>=1.24.3
pandas>=2.0.3
pillow>=10.2.0
langchain>=0.1.0
langchain-google-genai>=0.0.6
